Moving Clipboard related tests to a new class and combine the ClipboardTests class
