+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/ClipboardTests.cs
@@ -16,8 +16,16 @@ namespace System.Windows.Forms.Tests;
 
 [Collection("Sequential")]
 [CollectionDefinition("Sequential", DisableParallelization = true)]
-public partial class ClipboardTests
+public class ClipboardTests
 {
+    [WinFormsFact]
+    public void Clipboard_SetText_InvokeString_GetReturnsExpected()
+    {
+        Clipboard.SetText("text");
+        Assert.Equal("text", Clipboard.GetText());
+        Assert.True(Clipboard.ContainsText());
+    }
+
     [WinFormsFact]
     public void Clipboard_Clear_InvokeMultipleTimes_Success()
     {
